requirements.txt
git+https://github.com/Rapptz/discord.py
Pillow==9.4.0